升级框架与平台对接

+++ cases/AICardProject/case/marketingBusinessCenter/test_channelManagement.py
@@ -41,7 +41,7 @@ class test_ChannelManagement(TestCase):
     )
     def test_case_01_save_channel(self, data):
         # 新增渠道码
-        self.test_code = ["C00112"]
+        if not self._check_case(["C00112"]): return
         params = {
             'channel_name': data['channel_name'],
             'channel_input': data['channel_input']
@@ -54,7 +54,7 @@ class test_ChannelManagement(TestCase):
     )
     def test_case_02_edit_channel(self, data):
         # 编辑渠道码
-        self.test_code = ["C00113"]
+        if not self._check_case(["C00113"]): return
         params = {
             'channel_input': data['channel_input']
         }
@@ -63,25 +63,25 @@ class test_ChannelManagement(TestCase):
 
     def test_case_03_get_channel_code(self):
         # 获取渠道码
-        self.test_code = ["C00114"]
+        if not self._check_case(["C00114"]): return
         self.channelManagementLg.get_channel_code()
         time.sleep(3)
 
     def test_case_04_copy_link(self):
         # 复制链接
-        self.test_code = ["C00115"]
+        if not self._check_case(["C00115"]): return
         self.channelManagementLg.copy_link()
         time.sleep(3)
 
     def test_case_05_guest_detail(self):
         # 下载新客明细
-        self.test_code = ["C00116"]
+        if not self._check_case(["C00116"]): return
         self.channelManagementLg.guest_detail()
         time.sleep(5)
 
     def test_case_06_exposure_detail(self):
         # 下载曝光明细
-        self.test_code = ["C00117"]
+        if not self._check_case(["C00117"]): return
         self.channelManagementLg.exposure_detail()
         time.sleep(5)
 
@@ -90,7 +90,7 @@ class test_ChannelManagement(TestCase):
     )
     def test_case_07_search_channel(self, data):
         # 列表查询
-        self.test_code = ["C00118","C00119","C00120","C00121","C00122"]
+        if not self._check_case(["C00118","C00119","C00120","C00121","C00122"]): return
         params = {
             'channel_name': data['channel_name']
         }
@@ -99,27 +99,27 @@ class test_ChannelManagement(TestCase):
 
     def test_case_08_delete_channel(self):
         # 删除渠道码
-        self.test_code = ["C00123"]
+        if not self._check_case(["C00123"]): return
         self.channelManagementLg.delete_channel()
         time.sleep(3)
 
     def test_case_09_page_turning(self):
         # 点击下一页
-        self.test_code = ["C00124"]
+        if not self._check_case(["C00124"]): return
         self.channelManagementLg.page_turning()
         time.sleep(3)
 
 
     def test_case_10_single_path_code(self):
         # 获取单个路径码
-        self.test_code = ["C00125","C00126"]
+        if not self._check_case(["C00125","C00126"]): return
         self.channelManagementLg.switch_to_tab('获取路径码')
         self.channelManagementLg.single_path_code()
         time.sleep(3)
 
     def test_case_11_batch_code(self):
         # 获取批量码
-        self.test_code = ["C00128"]
+        if not self._check_case(["C00128"]): return
         self.channelManagementLg.switch_to_tab('获取批量码')
         self.channelManagementLg.batch_code()
         time.sleep(3)
